pyproject.toml
[build-system]
build-backend = "maturin"
requires = [ "maturin>=1,<2" ]

[project]
name = "drf-accelerator"
version = "0.1.0"
description = "A high-performance Rust-backed accelerator for Django Rest Framework."
readme = "README.md"
license = { text = "MIT" }

requires-python = ">=3.10"
classifiers = [
  "Framework :: Django",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
  "Programming Language :: Rust",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "djangorestframework>=3.12",
]

[dependency-groups]
dev = [
  "maturin>=1,<2",
  "pre-commit",
  "ruff",
]

[tool.maturin]
python-source = "drf_accelerator"
manifest-path = "drf_accelerator/Cargo.toml"
module-name = "drf_accelerator.drf_accelerator"

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = [ "examples" ]

lint.select = [ "E", "F", "I", "UP", "W" ]
lint.ignore = [  ]
lint.isort.known-first-party = [ "drf_accelerator" ]
